DepGen/requirements.txt
PyYAML==3.11
argparse==1.2.1
wsgiref==0.1.2
GitPython==0.3.2.RC1
async==0.6.1
gitdb==0.5.4
pycparser==2.10
smmap==0.8.2
